requirements.txt
blinker==1.9.0
click==8.2.0
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mutagen==1.47.0
psycopg2-binary==2.9.10
pymp3==0.2.0
python-dotenv==1.1.0
Werkzeug==3.1.3
